requirements.txt
catanatron>=3.2.1
catanatron_gym>=4.0.0
sb3-contrib>=2.2.1
stable-baselines3>=2.2.1
gymnasium>=0.29.1
torch>=2.0.0
pyyaml>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
tensorboard>=2.14.0
